changed some date things

+++ admin/all_restaurant.php
@@ -183,7 +183,7 @@ session_start();
 																								</div></td>
 																								
 																								<td>' . $rows['date'] . '</td>
-																									 <td><a href="delete_restaurant.php?res_del=' . $rows['rs_id'] . '" class="btn btn-danger btn-flat btn-addon btn-xs m-b-10"><i class="fa fa-trash-o" style="font-size:16px"></i></a> 
+                                                                                                     <td><a href="delete_restaurant.php?res_del=' . $rows['rs_id'] . '" class="btn btn-danger btn-flat btn-addon btn-xs m-b-10" onclick="return confirm(\'Are you sure you want to delete this restaurant?\')"><i class="fa fa-trash-o" style="font-size:16px"></i></a> 
 																									 <a href="update_restaurant.php?res_upd=' . $rows['rs_id'] . '" class="btn btn-info btn-flat btn-addon btn-sm m-b-10 m-l-5"><i class="fa fa-edit"></i></a>
 																									</td></tr>';
                                                 }
